Add main nav menu and separate add plan dialog component
