+++ components/InstanceControlBar.tsx
@@ -4,34 +4,21 @@ import { useEffect, useState } from 'react';
 
 import { useMutation, useSuspenseQuery } from '@apollo/client';
 import {
-  Alert,
-  AlertTitle,
   Box,
   Button,
-  CircularProgress,
   Container,
-  Dialog,
-  DialogActions,
-  DialogContent,
-  DialogTitle,
   FormControl,
-  FormHelperText,
-  FormLabel,
-  IconButton,
   InputLabel,
   MenuItem,
   Select,
   SelectChangeEvent,
-  Skeleton,
   Stack,
-  TextField,
-  Typography,
+  Link as MuiLink,
+  SxProps,
+  Theme,
 } from '@mui/material';
 import kebabCase from 'lodash/kebabCase';
-import startCase from 'lodash/startCase';
-import { ExclamationTriangle, FileEarmarkPlus, X } from 'react-bootstrap-icons';
 
-import { useUserProfile } from '@/hooks/use-user-profile';
 import { CREATE_NZC_FRAMEWORK_CONFIG } from '@/queries/framework/create-framework-config';
 import { GET_FRAMEWORK_CONFIGS } from '@/queries/framework/get-framework-config';
 import { useFrameworkInstanceStore } from '@/store/selected-framework-instance';
@@ -42,52 +29,10 @@ import {
   GetFrameworkConfigsQuery,
   LowHigh,
 } from '@/types/__generated__/graphql';
-import NumberInput from './NumberInput';
-
-type ClimateOption = 'warm' | 'cold';
-type RenewableMixOption = 'low' | 'high';
-type Option<T> = { label: string; value: T };
-
-const BASELINE_OPTIONS = [2018, 2019, 2020, 2021, 2022, 2023];
-const PANDEMIC_YEARS = [2020, 2021];
-
-const CLIMATE_OPTIONS: Option<ClimateOption>[] = [
-  { label: 'Warm (Above 12°C yearly average)', value: 'warm' },
-  { label: 'Cold (Below 12°C yearly average)', value: 'cold' },
-];
-
-const RENEWABLE_ELECTRICITY_OPTIONS: Option<RenewableMixOption>[] = [
-  { label: 'High (50-100% renewable & nuclear)', value: 'high' },
-  { label: 'Low (0-50% renewable & nuclear)', value: 'low' },
-];
-
-function getErrorMessage(error: Error) {
-  // Hacky way to check if the error is a duplicate key error. In future, our
-  // backend should return error codes to determine the type of 400 error.
-  if (
-    error.message.startsWith('Instance with identifier') &&
-    error.message.endsWith('already exists')
-  ) {
-    return 'A city plan with this name already exists. Please choose a different name.';
-  }
-
-  return `Error creating plan: ${error.message}`;
-}
-
-function isValid(data: Data, onlyStepOne = false): boolean {
-  const validFirstStep = !!(data.planName && data.baselineYear);
-
-  if (onlyStepOne) {
-    return validFirstStep;
-  }
-
-  return !!(
-    validFirstStep &&
-    data.population &&
-    data.renewableElectricityMix &&
-    data.climate
-  );
-}
+import { AddPlanDialog, NewPlanData } from './AddPlanDialog';
+import Link from 'next/link';
+import { usePathname } from 'next/navigation';
+import { areHistoricalYearsAvailable } from '@/utils/historical-data';
 
 function InstanceSelector({
   selectedInstanceId,
@@ -134,42 +79,26 @@ function InstanceSelector({
   );
 }
 
-type Data = {
-  planName: string;
-  baselineYear: number | '';
-  population: number | '';
-  climate: ClimateOption | null;
-  renewableElectricityMix: RenewableMixOption | null;
-};
-
-const INITIAL_DATA: Data = {
-  planName: '',
-  baselineYear: '',
-  population: '',
-  climate: null,
-  renewableElectricityMix: null,
-};
-
-function CityName() {
-  const { data, loading } = useUserProfile();
-
-  if (loading) {
-    return <Skeleton width={100} height={40} />;
-  }
-
-  const orgSlug = data?.me?.frameworkRoles?.[0]?.orgSlug;
-
-  if (!orgSlug) {
-    return null;
-  }
-
-  return <Typography color="text.secondary">{startCase(orgSlug)}</Typography>;
+function getNavStyles(isActive: boolean): SxProps<Theme> {
+  return {
+    color: 'primary.dark',
+    transition: 'border-width 0.1s',
+    borderColor: (theme) => theme.palette.primary.dark,
+    borderWidth: 0,
+    borderStyle: 'solid',
+    borderBottomWidth: isActive ? 2 : 0,
+    fontWeight: (theme) =>
+      isActive
+        ? theme.typography.fontWeightBold
+        : theme.typography.fontWeightRegular,
+    '&:hover': {
+      borderBottomWidth: 2,
+    },
+  };
 }
 
 export function InstanceControlBar() {
   const [isAddModalOpen, setIsAddModalOpen] = useState(false);
-  const [step, setStep] = useState(0);
-  const [data, setData] = useState<Data>(INITIAL_DATA);
 
   const { data: instanceData, error: instanceError } =
     useSuspenseQuery<GetFrameworkConfigsQuery>(GET_FRAMEWORK_CONFIGS);
@@ -179,58 +108,19 @@ export function InstanceControlBar() {
     CreateNzcFrameworkMutationVariables
   >(CREATE_NZC_FRAMEWORK_CONFIG);
 
-  async function handleSubmit(e: React.FormEvent<HTMLFormElement>) {
-    e.preventDefault();
-
-    if (data.planName && data.baselineYear) {
-      try {
-        const resp = await createFrameworkConfig({
-          variables: {
-            // TODO: Add climate, population and renewableElectricityMix to the mutation
-            frameworkId: 'nzc',
-            name: data.planName,
-            baselineYear: Number(data.baselineYear),
-            slug: kebabCase(data.planName),
-            population: Number(data.population),
-            renewableMix:
-              data.renewableElectricityMix === 'high'
-                ? LowHigh.High
-                : LowHigh.Low,
-            temperature: data.climate === 'warm' ? LowHigh.High : LowHigh.Low,
-          },
-        });
-
-        if (resp.data?.createNzcFrameworkConfig?.frameworkConfig?.id) {
-          const instance = resp.data.createNzcFrameworkConfig.frameworkConfig;
-
-          setInstance(
-            instance.id,
-            instance.organizationName ?? undefined,
-            instance.baselineYear
-          );
-        }
-
-        handleClose();
-      } catch (error) {
-        // TODO: Handle error
-        console.error('Error creating framework config:', error);
-      }
-    }
-  }
-
   const {
     data: selectedInstanceId,
     isDataInitialized: isInstanceStoreInitialized,
   } = useStore(useFrameworkInstanceStore, (state) => state.selectedInstance);
   const setInstance = useFrameworkInstanceStore((state) => state.setInstance);
 
+  const pathname = usePathname();
+
   useEffect(() => {
     if (
-      // If an instance is not selected, select the first one
       (isInstanceStoreInitialized &&
         !selectedInstanceId &&
         (instanceData.framework?.configs.length ?? 0 > 0)) ||
-      // If the selected instance is not in the list of instances, select the first one
       (isInstanceStoreInitialized &&
         !instanceData.framework?.configs.find(
           (config) => config.id === selectedInstanceId
@@ -254,22 +144,48 @@ export function InstanceControlBar() {
   ]);
 
   if (instanceError) {
-    // TODO: Error page
     return <div>Error: {instanceError.message}</div>;
   }
 
   const instanceConfigs = instanceData.framework?.configs ?? [];
   const hasMultipleInstances = instanceConfigs.length > 1;
+  const selectedInstance = instanceConfigs.find(
+    (instance) => instance.id === selectedInstanceId
+  );
 
-  function handleClose() {
-    setStep(0);
-    setIsAddModalOpen(false);
-    setData(INITIAL_DATA);
+  async function handleSubmit(data: NewPlanData) {
+    try {
+      const resp = await createFrameworkConfig({
+        variables: {
+          frameworkId: 'nzc',
+          name: data.planName,
+          baselineYear: Number(data.baselineYear),
+          slug: kebabCase(data.planName),
+          population: Number(data.population),
+          renewableMix:
+            data.renewableElectricityMix === 'high'
+              ? LowHigh.High
+              : LowHigh.Low,
+          temperature: data.climate === 'warm' ? LowHigh.High : LowHigh.Low,
+        },
+      });
+
+      if (resp.data?.createNzcFrameworkConfig?.frameworkConfig?.id) {
+        const instance = resp.data.createNzcFrameworkConfig.frameworkConfig;
+        setInstance(
+          instance.id,
+          instance.organizationName ?? undefined,
+          instance.baselineYear
+        );
+      }
+
+      setIsAddModalOpen(false);
+    } catch (error) {
+      console.error('Error creating framework config:', error);
+    }
   }
 
-  function handleChange(field: keyof Data, value: Data[typeof field]) {
-    setData((prev) => ({ ...prev, [field]: value }));
-  }
+  const isActive = (path: string) => pathname === path;
 
   return (
     <>
@@ -290,7 +206,29 @@ export function InstanceControlBar() {
             alignItems: 'center',
           }}
         >
-          <CityName />
+          {!!selectedInstance && (
+            <Stack direction="row" spacing={2}>
+              <MuiLink
+                href="/"
+                component={Link}
+                underline="none"
+                sx={getNavStyles(isActive('/'))}
+              >
+                Overview
+              </MuiLink>
+
+              {areHistoricalYearsAvailable(selectedInstance.baselineYear) && (
+                <MuiLink
+                  href="/additional-data"
+                  component={Link}
+                  underline="none"
+                  sx={getNavStyles(isActive('/additional-data'))}
+                >
+                  Additional historical data
+                </MuiLink>
+              )}
+            </Stack>
+          )}
 
           <Stack
             direction="row"
@@ -311,239 +249,13 @@ export function InstanceControlBar() {
         </Container>
       </Box>
 
-      <Dialog
+      <AddPlanDialog
         open={isAddModalOpen}
-        onClose={handleClose}
-        maxWidth="xs"
-        fullWidth
-      >
-        <DialogTitle sx={{ pb: 1 }}>
-          <Stack
-            direction="row"
-            justifyContent="space-between"
-            alignItems="center"
-            spacing={1}
-          >
-            <Box sx={{ fontSize: 0 }}>
-              <FileEarmarkPlus size={24} />
-            </Box>
-            <Box component="span" flex="1">
-              <div>Create new plan</div>
-            </Box>
-            <IconButton aria-label="close" onClick={handleClose}>
-              <X size={24} />
-            </IconButton>
-          </Stack>
-        </DialogTitle>
-        <form noValidate autoComplete="off" onSubmit={handleSubmit}>
-          <DialogContent sx={{ px: 3, pt: 0 }}>
-            <Box sx={{ mb: 4 }}>
-              {step === 0 && (
-                <Typography variant="body1" color="text.secondary">
-                  Create a new climate action plan to track and manage your
-                  city's emissions data. Create multiple plans to envision
-                  various scenarios.
-                </Typography>
-              )}
-
-              {step === 1 && (
-                <Typography variant="body1" color="text.secondary">
-                  {data.planName} ({data.baselineYear})
-                </Typography>
-              )}
-            </Box>
-
-            {step === 0 && (
-              <Stack spacing={2}>
-                <TextField
-                  required
-                  label="Plan or city name"
-                  value={data.planName}
-                  onChange={(e) => handleChange('planName', e.target.value)}
-                />
-                <FormControl required>
-                  <InputLabel id="baseline-select">Baseline year</InputLabel>
-                  <Select
-                    label="Baseline year"
-                    labelId="baseline-select"
-                    id="baseline-select-component"
-                    renderValue={(value) => value}
-                    value={data.baselineYear}
-                    onChange={(e) =>
-                      handleChange('baselineYear', Number(e.target.value))
-                    }
-                  >
-                    {BASELINE_OPTIONS.map((year) => (
-                      <MenuItem key={year} value={year}>
-                        <Stack spacing={0.5}>
-                          <Typography>{year}</Typography>
-                          {PANDEMIC_YEARS.includes(year) && (
-                            <Stack
-                              sx={{ color: 'text.secondary' }}
-                              direction="row"
-                              spacing={0.5}
-                              justifyContent="center"
-                            >
-                              <Box sx={{ color: 'warning.main' }}>
-                                <ExclamationTriangle size={16} />
-                              </Box>
-                              <Typography variant="caption">
-                                COVID-19 may have skewed data for this year
-                              </Typography>
-                            </Stack>
-                          )}
-                        </Stack>
-                      </MenuItem>
-                    ))}
-                  </Select>
-                  {data.baselineYear &&
-                  PANDEMIC_YEARS.includes(data.baselineYear) ? (
-                    <FormHelperText
-                      component={Stack}
-                      sx={{ color: 'warning.dark', pt: 1 }}
-                      direction="row"
-                      spacing={1}
-                      justifyContent="center"
-                    >
-                      <Box>
-                        <ExclamationTriangle size={18} />
-                      </Box>
-                      <Typography variant="caption">
-                        COVID-19 may have skewed data for this year. Consider
-                        another baseline year for more typical results.
-                      </Typography>
-                    </FormHelperText>
-                  ) : (
-                    <FormHelperText>
-                      The baseline year is the reference point for measuring
-                      future emission reductions. You must provide your city's
-                      operational and statistical data for this specific year.
-                    </FormHelperText>
-                  )}
-                </FormControl>
-              </Stack>
-            )}
-
-            {step === 1 && (
-              <Stack spacing={2}>
-                <FormControl required>
-                  <FormLabel id="population-input" sx={{ mb: 0.5 }}>
-                    <Typography component="span" variant="body2">
-                      What's your city's population?
-                    </Typography>
-                  </FormLabel>
-                  <NumberInput
-                    aria-labelledby="population-input"
-                    inputProps={{
-                      allowNegative: false,
-                      min: 0,
-                      max: 50000000,
-                    }}
-                    value={data.population}
-                    onValueChange={(values) =>
-                      handleChange('population', values.floatValue ?? '')
-                    }
-                  />
-                </FormControl>
-
-                <FormControl required>
-                  <FormLabel id="climate-select" sx={{ mb: 0.5 }}>
-                    <Typography component="span" variant="body2">
-                      Is your city warm or cold?
-                    </Typography>
-                  </FormLabel>
-                  <Select
-                    hiddenLabel
-                    labelId="climate-select"
-                    id="climate-select-component"
-                    value={data.climate ?? ''}
-                    onChange={(e) => handleChange('climate', e.target.value)}
-                  >
-                    {CLIMATE_OPTIONS.map(({ value, label }) => (
-                      <MenuItem key={value} value={value}>
-                        {label}
-                      </MenuItem>
-                    ))}
-                  </Select>
-                </FormControl>
-
-                <FormControl required>
-                  <FormLabel id="electricity-select" sx={{ mb: 0.5 }}>
-                    <Typography component="span" variant="body2">
-                      What's the percentage of renewable plus nuclear energy in
-                      your electricity mix?
-                    </Typography>
-                  </FormLabel>
-                  <Select
-                    hiddenLabel
-                    labelId="electricity-select"
-                    id="electricity-select-component"
-                    value={data.renewableElectricityMix ?? ''}
-                    onChange={(e) =>
-                      handleChange('renewableElectricityMix', e.target.value)
-                    }
-                  >
-                    {RENEWABLE_ELECTRICITY_OPTIONS.map(({ value, label }) => (
-                      <MenuItem key={value} value={value}>
-                        {label}
-                      </MenuItem>
-                    ))}
-                  </Select>
-                </FormControl>
-              </Stack>
-            )}
-
-            {error && (
-              <Alert severity="error" sx={{ mt: 2 }}>
-                <AlertTitle>{getErrorMessage(error)}</AlertTitle>
-              </Alert>
-            )}
-          </DialogContent>
-
-          <DialogActions
-            sx={{
-              justifyContent: 'flex-end',
-              px: 3,
-              pb: 3,
-            }}
-          >
-            {step === 0 && (
-              <>
-                <Button variant="text" onClick={handleClose}>
-                  Cancel
-                </Button>
-                <Button
-                  variant="contained"
-                  onClick={() => setStep(1)}
-                  disabled={!isValid(data, true)}
-                >
-                  Next
-                </Button>
-              </>
-            )}
-
-            {step === 1 && (
-              <>
-                <Button variant="text" onClick={() => setStep(0)}>
-                  Previous
-                </Button>
-                <Button
-                  variant="contained"
-                  type="submit"
-                  disabled={loading || !isValid(data)}
-                  endIcon={
-                    loading ? (
-                      <CircularProgress color="inherit" size={20} />
-                    ) : null
-                  }
-                >
-                  {loading ? 'Adding...' : 'Add plan'}
-                </Button>
-              </>
-            )}
-          </DialogActions>
-        </form>
-      </Dialog>
+        onClose={() => setIsAddModalOpen(false)}
+        onSubmit={handleSubmit}
+        loading={loading}
+        error={error}
+      />
     </>
   );
 }
